Fixes animation issue where the pull to refresh header would get out of place
When pulling the table view, we properly position the pull to refresh header so that it seems as if it's being pulled out of the navigation bar. However, when it's animating down, it gets into a state where the shape of the view will not change as the animation is still in progress. Therefore we only begin observing the contentInset changes again once the animation is done, avoiding the problem all together.